Fix handling of file names with spaces; update all stats files
We had a problem with the attributes file; file names with spaces
were being incorrectly processed. Some files now need to have
tab-separated fields, and the Makefile `gawk` scripts needed
to be updated. This led to a large number of file changes.

Also simplified the Makefile by creating `fortran-file-patterns.txt`
file to replace invocations of `find` with `grep` on `all-files.txt`.

Also added a `fortran-exceptions.txt` file which contains the names
of files that /appear/ to be Fortran (by name) but aren't (such as
the various `Makefile.f90` files).